Move tests for unary functions into a separate file
