+++ tests/test_dt_expr.py
@@ -53,246 +53,6 @@ dt_obj = [[dt, pytest, random, f, dt_int, None],
 
 
 #-------------------------------------------------------------------------------
-# Unary bitwise NOT (__invert__)
-#-------------------------------------------------------------------------------
-
-def inv(t):
-    if t is None: return None
-    if isinstance(t, bool): return not t
-    return ~t
-
-
-@pytest.mark.parametrize("src", dt_bool + dt_int)
-def test_dt_invert(src):
-    dt0 = dt.Frame(src)
-    df2 = dt0[:, ~f[0]]
-    frame_integrity_check(df2)
-    assert df2.stypes == dt0.stypes
-    assert df2.to_list() == [[inv(x) for x in src]]
-
-
-@pytest.mark.parametrize("src", dt_float)
-def test_dt_invert_invalid(src):
-    dt0 = dt.Frame(src)
-    with pytest.raises(TypeError) as e:
-        noop(dt0[:, ~f[0]])
-    assert ("Cannot apply unary `operator ~` to a column with stype `%s`"
-            % dt0.stypes[0].name == str(e.value))
-
-
-
-#-------------------------------------------------------------------------------
-# Unary minus (__neg__)
-#-------------------------------------------------------------------------------
-
-def neg(t):
-    if t is None: return None
-    return -t
-
-
-@pytest.mark.parametrize("src", dt_int + dt_float + dt_bool)
-def test_dt_neg(src):
-    dt0 = dt.Frame(src)
-    dtr = dt0[:, -f[0]]
-    frame_integrity_check(dtr)
-    assert dtr.stypes == ((dt.int8,) if dt0.stypes[0] == dt.bool8 else
-                          dt0.stypes)
-    assert_equals(dtr, dt.Frame([neg(x) for x in src]))
-
-
-@pytest.mark.parametrize("src", dt_str)
-def test_dt_neg_invalid(src):
-    dt0 = dt.Frame(src)
-    with pytest.raises(TypeError) as e:
-        noop(dt0[:, -f[0]])
-    assert ("Cannot apply unary `operator -` to a column with stype `%s`"
-            % dt0.stypes[0].name == str(e.value))
-
-
-
-#-------------------------------------------------------------------------------
-# Unary plus (__pos__)
-#-------------------------------------------------------------------------------
-
-@pytest.mark.parametrize("src", dt_int + dt_float + dt_bool)
-def test_dt_pos(src):
-    dt0 = dt.Frame(src)
-    dtr = dt0[:, +f[0]]
-    frame_integrity_check(dtr)
-    stype0 = dt0.stypes[0]
-    if stype0 == stype.bool8:
-        stype0 = stype.int8
-    assert dtr.stypes == (stype0,)
-    assert list_equals(dtr.to_list()[0], list(src))
-
-
-@pytest.mark.parametrize("src", dt_str)
-def test_dt_pos_invalid(src):
-    dt0 = dt.Frame(src)
-    with pytest.raises(TypeError) as e:
-        noop(dt0[:, +f[0]])
-    assert ("Cannot apply unary `operator +` to a column with stype `%s`"
-            % dt0.stypes[0].name == str(e.value))
-
-
-
-#-------------------------------------------------------------------------------
-# isna()
-#-------------------------------------------------------------------------------
-
-@pytest.mark.parametrize("src", dt_bool + dt_int + dt_float + dt_str)
-def test_dt_isna(src):
-    from datatable import isna
-    dt0 = dt.Frame(src)
-    dt1 = dt0[:, isna(f[0])]
-    frame_integrity_check(dt1)
-    assert dt1.stypes == (stype.bool8,)
-    pyans = [x is None for x in src]
-    assert dt1.to_list()[0] == pyans
-
-
-def test_dt_isna2():
-    from datatable import isna
-    from math import nan
-    dt0 = dt.Frame(A=[1, None, 2, 5, None, 3.6, nan, -4.899])
-    dt1 = dt0[~isna(f.A), :]
-    assert dt1.stypes == dt0.stypes
-    assert dt1.names == dt0.names
-    assert dt1.to_list() == [[1.0, 2.0, 5.0, 3.6, -4.899]]
-
-
-
-
-#-------------------------------------------------------------------------------
-# abs()
-#-------------------------------------------------------------------------------
-
-def test_abs():
-    from datatable import abs
-    assert abs(1) == 1
-    assert abs(-5) == 5
-    assert abs(-2.5e12) == 2.5e12
-    assert abs(None) is None
-
-
-@pytest.mark.parametrize("src", dt_int + dt_float)
-def test_abs_srcs(src):
-    dt0 = dt.Frame(src)
-    dt1 = dt.abs(dt0)
-    frame_integrity_check(dt1)
-    assert dt0.stypes == dt1.stypes
-    pyans = [None if x is None else abs(x) for x in src]
-    assert dt1.to_list()[0] == pyans
-
-
-def test_abs_all_stypes():
-    from datatable import abs
-    src = [[-127, -5, -1, 0, 2, 127],
-           [-32767, -299, -7, 32767, 12, -543],
-           [-2147483647, -1000, 3, -589, 2147483647, 0],
-           [-2**63 + 1, 2**63 - 1, 0, -2**32, 2**32, -793]]
-    dt0 = dt.Frame(src, stypes=[dt.int8, dt.int16, dt.int32, dt.int64])
-    dt1 = dt0[:, [abs(f[i]) for i in range(4)]]
-    frame_integrity_check(dt1)
-    assert dt1.to_list() == [[abs(x) for x in col] for col in src]
-
-
-
-#-------------------------------------------------------------------------------
-# exp()
-#-------------------------------------------------------------------------------
-
-def test_exp():
-    from datatable import exp
-    assert exp(0) == math.exp(0)
-    assert exp(1) == math.exp(1)
-    assert exp(-2.5e12) == math.exp(-2.5e12)
-    assert exp(12345678) == math.inf
-    assert exp(None) is None
-
-
-@pytest.mark.parametrize("src", dt_int + dt_float)
-def test_exp_srcs(src):
-    from math import exp, inf
-    dt0 = dt.Frame(src)
-    dt1 = dt.exp(dt0)
-    frame_integrity_check(dt1)
-    assert all([st == stype.float64 for st in dt1.stypes])
-    pyans = []
-    for x in src:
-        if x is None:
-            pyans.append(None)
-        else:
-            try:
-                pyans.append(exp(x))
-            except OverflowError:
-                pyans.append(inf)
-    assert dt1.to_list()[0] == pyans
-
-
-def test_exp_all_stypes():
-    from datatable import exp
-    src = [[-127, -5, -1, 0, 2, 127],
-           [-32767, -299, -7, 32767, 12, -543],
-           [-2147483647, -1000, 3, -589, 2147483647, 0],
-           [-2 ** 63 + 1, 2 ** 63 - 1, 0, -2 ** 32, 2 ** 32, -793]]
-    dt0 = dt.Frame(src, stypes=[dt.int8, dt.int16, dt.int32, dt.int64])
-    dt1 = dt0[:, [exp(f[i]) for i in range(4)]]
-    frame_integrity_check(dt1)
-    pyans = []
-    for col in src:
-        l = []
-        for x in col:
-            if x is None:
-                l.append(None)
-            else:
-                try:
-                    l.append(math.exp(x))
-                except OverflowError:
-                    l.append(math.inf)
-        pyans.append(l)
-    assert dt1.to_list() == pyans
-
-
-
-
-#-------------------------------------------------------------------------------
-# log(), log10()
-#-------------------------------------------------------------------------------
-
-@pytest.mark.parametrize("fn", ["log", "log10"])
-def test_log(fn):
-    dtlog = getattr(dt, fn)
-    mathlog = getattr(math, fn)
-    assert dtlog(None) is None
-    assert dtlog(-1) is None
-    assert dtlog(0) == -math.inf
-    assert dtlog(1) == 0.0
-    assert dtlog(5) == mathlog(5)
-    assert dtlog(10) == mathlog(10)
-    assert dtlog(793.54) == mathlog(793.54)
-    assert dtlog(math.inf) == math.inf
-
-
-@pytest.mark.parametrize("src", dt_int + dt_float)
-@pytest.mark.parametrize("fn", ["log", "log10"])
-def test_log_srcs(src, fn):
-    dtlog = getattr(dt, fn)
-    mathlog = getattr(math, fn)
-    dt0 = dt.Frame(src)
-    dt1 = dtlog(dt0)
-    frame_integrity_check(dt1)
-    assert all([st == stype.float64 for st in dt1.stypes])
-    pyans = [None if x is None or x < 0 else
-             -math.inf if x == 0 else
-             mathlog(x)
-             for x in src]
-    assert dt1.to_list()[0] == pyans
-
-
-
-
-#-------------------------------------------------------------------------------
 # logical ops
 #-------------------------------------------------------------------------------
